requirement.txt
selenium
django
djangorestframework